\name{communities}
\alias{communities}
\title{List communities}
\description{
  Returns the community labels associated with a phylo4com object.
}
\usage{
  communities(x)
}
\arguments{
  \item{x}{a phylo4com object}
}
\value{
  Vector of community labels.
}
\author{Jim Regetz (regetz@nceas.ucsb.edu)}
\examples{
  communities(weeds)
}
