\name{ecoPD-package}
\alias{ecoPD-package}
\alias{ecoPD}
\docType{package}
\title{
Ecologically-informed phylodiversity metrics in R
}
\description{
This package is intended to provide a novel set of metrics and functions
for combining ecological and phylogenetic information.
}
\details{
\tabular{ll}{
Package: \tab ecoPD\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2008-09-03\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
}
\author{
  Authors: Jim Regetz, Marc Cadotte, and Jonathan Davies, with support
  from the National Center for Ecological Analysis and Synthesis (NCEAS)
}
\references{
  Cadotte, M., Davies, J., Regetz, J., Kembel, S., Cleland, E., and
  Oakley, T. (In press) Phylogenetic diversity metrics for ecological
  communities: integrating species richness, abundance and evolutionary
  history. \emph{Ecology Letters}.
}
\keyword{ package }
