\name{LookupTL}
\alias{LookupTL}
\docType{data}
\title{Family-based lookup table for minimum tip lengths}
\description{
  TODO
}
\format{
  A data frame with information on 404 families, referenced by rowname:
  \describe{
    \item{\code{age}}{age of family}
    \item{\code{species}}{species richness of family}
    \item{\code{minTL}}{minimum tip length}
    \item{\code{apg2supertree}}{notes on association between angiosperm
      phylogeny group and supertree tip label}
    \item{\code{supertree}}{supertree tip label}
  }
}
\details{
  TODO
}
\source{
  TODO
}
\references{
  TODO
}
\keyword{datasets}
