\name{weeds}
\alias{weeds}
\docType{data}
\title{Weed phylogeny and community data}
\description{
  Phylogeny and abundance data for a small set of weedy plant species
  occurring in three hypothetical plots.
}
\format{
  A list \code{phylo4d} objects, each with abundance data for tip taxa.
}
\details{
  This data set includes five weedy annual plant species commonly found
  across North America. Abundance values represent fabricated data for
  three hypothetical field sites (named A, B, and C). The phylogeny
  component of the dataset is real, and was estimated from four genes:
  rbcL, MatK, ITS1, and 5.8s.
}
\references{
  Cadotte, M., Davies, J., Regetz, J., Kembel, S., Cleland, E., and
  Oakley, T. (In press) Phylogenetic diversity metrics for ecological
  communities: integrating species richness, abundance and evolutionary
  history. \emph{Ecology Letters}.
}
\keyword{datasets}
